CHG: run communication test in .c; fixes included
